+++ src/main/java/br/com/pi/projeto_RD/model/dto/FilialDTO.java
@@ -13,7 +13,7 @@ import java.util.List;
 @Data
 public class FilialDTO {
 
-    private Integer cd_filial;
+    private Long cd_filial;
     private String nm_filial;
     private String nr_cpnj;
     private String nr_telefone;
